[ADD] Add new hook point for ECCUBE version >= 3.0.9 (done).
[ADD] Add new event (shopping confirm init).

[CHANGE] Change to new event.

[ADD] Add new event (shopping index).

[CHANGE] Change from twig event to response event.
